search: simplify deduper
`Deduper` had a fairly expensive logic to keep track of deduped results,
but none of the call sites read the results. This simplifies the
implementation without touching the call sites.

Additionally, I remove `dedupingStream` which used `Deduper` internally,
but is not used anywhere itself.

Test plan:
- updated unit test
- CI
