+++ internal/search/result/BUILD.bazel
@@ -7,7 +7,6 @@ go_library(
         "commit.go",
         "commit_diff.go",
         "commit_json.go",
-        "deduper.go",
         "file.go",
         "highlight.go",
         "match.go",
@@ -43,7 +42,6 @@ go_test(
         "commit_diff_test.go",
         "commit_json_test.go",
         "commit_test.go",
-        "deduper_test.go",
         "file_test.go",
         "match_test.go",
         "merger_test.go",
